backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
greenlet==3.3.1
alembic==1.14.0

# Redis and Celery
redis==5.2.0
celery[redis]==5.4.0

# HTTP client
httpx==0.28.1

# Scientific computing
numpy==2.2.1

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# CORS
python-multipart==0.0.18
